Task4/sdk-ds/requirements.txt
akeyless==3.5.3
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.0
